[NFC] regen code after 80ffbb7e8960 and 4f174fb7
If you need to rebase as a result of the churn, all that should be
needed is to run the generate target on "theirs".

n.b. In addition to running the generate to clang-format the code, the
license headers were also manually tweaked to ensure SPDX license
identifiers in comment headers are alone on a line as required by the
spec. clang-format was not clever enough to do this (nor should it be)